backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
Werkzeug==3.0.1
python-socketio==5.11.0
eventlet==0.33.3
dnspython==2.2.0
requests==2.31.0
Pillow==10.2.0
openai==1.12.0
authlib==1.3.0
cryptography==41.0.7
openpyxl==3.1.2
reportlab==4.0.7
python-docx==1.1.0
PyPDF2==3.0.1
pyotp==2.9.0
qrcode==7.4.2
celery==5.3.4
redis==5.0.1
whoosh==2.7.4
flask-limiter==3.5.0
prometheus-client==0.19.0
boto3==1.42.0
minio==7.1.2
hvac==2.4.0
livekit-api==0.3.0
y-py==0.6.2
ypy-websocket==0.12.1
agora-token-builder==1.0.0
firebase-functions
firebase-admin
Flask-Mail==0.9.1
gunicorn==21.2.0
